+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Changes in v0.2.7
 =================
 
 - Added ``HotQueue.bulk`` context manager for enqueuing items in bulk
+- Added support for completely disabling serialization
 
 Changes in v0.2.6
 =================